改进登出成功后的处理机制

+++ tnxjeex-cas/tnxjeex-cas-server/src/main/java/org/truenewx/tnxjeex/cas/server/ticket/TicketManagerImpl.java
@@ -47,49 +47,57 @@ public class TicketManagerImpl implements TicketManager, HttpSessionListener {
     public void createTicketGrantingTicket(HttpServletRequest request,
             HttpServletResponse response) {
         String sessionId = request.getSession().getId();
-        String ticketGrantingTicket = getTicketGrantingTicket(sessionId);
+        String ticketGrantingTicket = createTicketGrantingTicket(sessionId);
         int maxAge = (int) this.serverProperties.getServlet().getSession().getTimeout().toSeconds();
         WebUtil.addCookie(request, response, COOKIE_TGT, ticketGrantingTicket, maxAge);
     }
 
-    private String getTicketGrantingTicket(String sessionId) {
+    private String createTicketGrantingTicket(String sessionId) {
         return TICKET_GRANTING_TICKET_PREFIX + EncryptUtil.encryptByMd5_16(sessionId);
+    }
+
+    private String getTicketGrantingTicket(HttpServletRequest request, boolean create) {
+        String ticket = WebUtil.getCookieValue(request, COOKIE_TGT);
+        if (ticket == null && create) {
+            ticket = createTicketGrantingTicket(request.getSession().getId());
+        }
+        return ticket;
     }
 
     @Override
     public boolean validateTicketGrantingTicket(HttpServletRequest request) {
-        String ticketGrantingTicket = WebUtil.getCookieValue(request, COOKIE_TGT);
+        String ticketGrantingTicket = getTicketGrantingTicket(request, false);
         String sessionId = request.getSession().getId();
         return StringUtils.isNotBlank(ticketGrantingTicket)
-                && ticketGrantingTicket.equals(getTicketGrantingTicket(sessionId));
+                && ticketGrantingTicket.equals(createTicketGrantingTicket(sessionId));
     }
 
     @Override
     public Map<String, String> deleteServiceTickets(HttpServletRequest request) {
         Map<String, String> mapping = new HashMap<>();
-        String sessionId = request.getSession().getId();
-        String ticketGrantingTicket = getTicketGrantingTicket(sessionId);
-        List<ServiceTicket> tickets = this.serviceTicketRepo.deleteByTicketGrantingTicket(ticketGrantingTicket);
-        tickets.forEach(ticket -> {
-            mapping.put(ticket.getService(), ticket.getId());
-        });
+        String ticketGrantingTicket = getTicketGrantingTicket(request, false);
+        if (ticketGrantingTicket != null) {
+            List<ServiceTicket> tickets = this.serviceTicketRepo.deleteByTicketGrantingTicket(ticketGrantingTicket);
+            tickets.forEach(ticket -> {
+                mapping.put(ticket.getService(), ticket.getId());
+            });
+        }
         return mapping;
     }
 
     // 用户登录或登出CAS服务器成功后调用，以获取目标服务的票据
     @Override
     @WriteTransactional
-    public String getServiceTicket(HttpServletRequest request, String service, boolean create) {
-        String sessionId = request.getSession().getId();
-        String ticketGrantingTicket = getTicketGrantingTicket(sessionId);
+    public String getServiceTicket(HttpServletRequest request, String service) {
+        String ticketGrantingTicket = getTicketGrantingTicket(request, true);
         ServiceTicket ticket = this.serviceTicketRepo.findByTicketGrantingTicketAndService(ticketGrantingTicket, service);
         if (ticket != null && ticket.getExpiredTime().before(new Date())) { // 已过期的先删除，再视为null
             this.serviceTicketRepo.deleteById(ticket.getId());
             ticket = null;
         }
-        if (ticket == null && create) { // 不存在且需要创建则创建新的
+        if (ticket == null) { // 不存在则创建新的
             Date now = new Date();
-            String text = sessionId + Strings.MINUS + service + Strings.MINUS + now.getTime();
+            String text = ticketGrantingTicket + Strings.MINUS + service + Strings.MINUS + now.getTime();
             String ticketId = SERVICE_TICKET_PREFIX + EncryptUtil.encryptByMd5_16(text);
             long timeout = this.serverProperties.getServlet().getSession().getTimeout().toMillis();
             Date expiredTime = new Date(now.getTime() + timeout);
@@ -101,7 +109,7 @@ public class TicketManagerImpl implements TicketManager, HttpSessionListener {
             ticket.setExpiredTime(expiredTime);
             this.serviceTicketRepo.save(ticket);
         }
-        return ticket == null ? null : ticket.getId();
+        return ticket.getId();
     }
 
     // 用户访问业务服务，由业务服务校验票据有效性时调用
@@ -130,7 +138,7 @@ public class TicketManagerImpl implements TicketManager, HttpSessionListener {
     @Override
     public void sessionDestroyed(HttpSessionEvent event) {
         String sessionId = event.getSession().getId();
-        String ticketGrantingTicket = getTicketGrantingTicket(sessionId);
+        String ticketGrantingTicket = createTicketGrantingTicket(sessionId);
         List<ServiceTicket> tickets = this.serviceTicketRepo.deleteByTicketGrantingTicket(ticketGrantingTicket);
         tickets.forEach(ticket -> {
             LogUtil.info(getClass(),
